move source to light
