+++ caustic/light/base.py
