+++ caustic/light/pixelated.py
